booksapp/requirements.txt
asgiref==3.4.1
attrs==21.4.0
backports.zoneinfo==0.2.1
Django==4.0.1
djangorestframework==3.13.1
Faker==11.3.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2